+++ src/enums/deviceType.ts
@@ -40,6 +40,7 @@ export enum DeviceType {
   ZigbeeInnr142C = 226,
   ZigbeeSodaHandle = 227,
   ZigbeeTuyaMotion = 228,
+  ZigbeeAqaraPresence = 229,
   JsEnergyManager = 301,
   RoomScene = 401,
   ShellyTrv = 402,